// src/api/generated/models/materialRequestSummary.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { MaterialRequestSummaryCreatedByUserId } from './materialRequestSummaryCreatedByUserId';
import type { MaterialRequestSummaryCreatedByUserName } from './materialRequestSummaryCreatedByUserName';
import type { MaterialRequestSummaryAssigneeUserId } from './materialRequestSummaryAssigneeUserId';
import type { MaterialRequestSummaryAssigneeUserName } from './materialRequestSummaryAssigneeUserName';
import type { MaterialRequestSummaryReference } from './materialRequestSummaryReference';
import type { MaterialRequestSummaryStatus } from './materialRequestSummaryStatus';
import type { MaterialRequestSummaryPriority } from './materialRequestSummaryPriority';
import type { MaterialRequestSummaryDueDate } from './materialRequestSummaryDueDate';
import type { MaterialRequestSummarySubmittedAt } from './materialRequestSummarySubmittedAt';

export interface MaterialRequestSummary {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  companyId: string;
  createdByUserId: MaterialRequestSummaryCreatedByUserId;
  createdByUserName: MaterialRequestSummaryCreatedByUserName;
  assigneeUserId: MaterialRequestSummaryAssigneeUserId;
  assigneeUserName: MaterialRequestSummaryAssigneeUserName;
  title: string;
  reference: MaterialRequestSummaryReference;
  status: MaterialRequestSummaryStatus;
  priority: MaterialRequestSummaryPriority;
  dueDate: MaterialRequestSummaryDueDate;
  submittedAt: MaterialRequestSummarySubmittedAt;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  updatedAt: string;
}